Don't unbind intermediate keys in targets-setup
This allows targets-setup to be run without unbinding previous next,
last, and remote text objects. "n", "l", and "r" do not have text
objects by default in evil. If user the changes any of these keys, it is
now up to them to unbind the new key(s) if necessary.