Add improved method for inner vs. outer text objects
Using two things is not really suitable because the range for the inner
thing could be after the point. For example, for a string text object
with the point just before the string (i.e. \"string"), the outer text
object would work, but the inner text object would not because the
selection would be after the point (making it seem as if there is not a
thing at the point).

Targets will now shrink the range afterwards to prevent this issue.

Addresses #16.